requirements.txt
fastapi==0.82.0
uvicorn==0.18.2
qrcode 
Pillow